+++ src/modules/book/book.controller.spec.ts
@@ -2,9 +2,27 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { BookController } from './book.controller';
 import { BookService } from './book.service';
 import { CreateBookService } from './services/create-book.service';
+import { FindBookByIdService } from './services/find-book-by-id.service';
 import { FindAllBooksService } from './services/find-all-books.service';
+import { UpdateBookByIdService } from './services/update-book.service';
+import { DeleteBookService } from './services/delete-book.service';
+import { CanActivate } from '@nestjs/common';
+import { ApiKeyAuthGuard } from 'src/common/guards/api-key.guard';
 
-const mockBookRepository = { find: jest.fn(), save: jest.fn() };
+const mockBookRepository = {
+  create: jest.fn(),
+  findAll: jest.fn(),
+  findOne: jest.fn(),
+  update: jest.fn(),
+  delete: jest.fn(),
+};
+
+// Mock ApiKeyAuthGuard
+const mockApiKeyAuthGuard: CanActivate = {
+  canActivate: () => {
+    return true; // Allow access to the guarded routes during testing
+  },
+};
 
 describe('BookController', () => {
   let controller: BookController;
@@ -17,8 +35,14 @@ describe('BookController', () => {
         CreateBookService,
         FindAllBooksService,
         { provide: 'BookRepository', useValue: mockBookRepository },
+        FindBookByIdService,
+        UpdateBookByIdService,
+        DeleteBookService,
       ],
-    }).compile();
+    })
+      .overrideGuard(ApiKeyAuthGuard) // Override the guard
+      .useValue(mockApiKeyAuthGuard)
+      .compile();
 
     controller = module.get<BookController>(BookController);
   });