+++ src/modules/book/book.service.spec.ts
@@ -1,7 +1,10 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { BookService } from './book.service';
 import { CreateBookService } from './services/create-book.service';
+import { FindBookByIdService } from './services/find-book-by-id.service';
 import { FindAllBooksService } from './services/find-all-books.service';
+import { UpdateBookByIdService } from './services/update-book.service';
+import { DeleteBookService } from './services/delete-book.service';
 
 const mockBookRepository = {
   // You can add mock implementations for any methods you'll be using.
@@ -22,6 +25,9 @@ describe('BookService', () => {
           provide: 'BookRepository',
           useValue: mockBookRepository, // Use the mock for BookRepository
         },
+        FindBookByIdService,
+        UpdateBookByIdService,
+        DeleteBookService,
       ],
     }).compile();