+++ CHANGELOG.md
@@ -2,6 +2,75 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.27](https://github.com/esgiraldop/library-nest-js/compare/v0.0.25...v0.0.27) (2024-10-02)
+
+
+### Features
+
+* Implement x-api-key guard ([c4fd28b](https://github.com/esgiraldop/library-nest-js/commit/c4fd28bc772f8539ee9ea64b2582d8c3a320afc8))
+
+### [0.0.26](https://github.com/esgiraldop/library-nest-js/compare/v0.0.25...v0.0.26) (2024-10-01)
+
+
+### Features
+
+* Implement x-api-key guard ([c4fd28b](https://github.com/esgiraldop/library-nest-js/commit/c4fd28bc772f8539ee9ea64b2582d8c3a320afc8))
+
+### [0.0.25](https://github.com/esgiraldop/library-nest-js/compare/v0.0.24...v0.0.25) (2024-10-01)
+
+### [0.0.24](https://github.com/esgiraldop/library-nest-js/compare/v0.0.23...v0.0.24) (2024-10-01)
+
+### [0.0.23](https://github.com/esgiraldop/library-nest-js/compare/v0.0.22...v0.0.23) (2024-10-01)
+
+### [0.0.22](https://github.com/esgiraldop/library-nest-js/compare/v0.0.21...v0.0.22) (2024-10-01)
+
+### [0.0.21](https://github.com/esgiraldop/library-nest-js/compare/v0.0.20...v0.0.21) (2024-10-01)
+
+
+### Features
+
+* create delete service for books ([6fd5018](https://github.com/esgiraldop/library-nest-js/commit/6fd5018f8ed6eb14bc7e5028777a8ea59af01897))
+
+### [0.0.20](https://github.com/esgiraldop/library-nest-js/compare/v0.0.19...v0.0.20) (2024-10-01)
+
+
+### Features
+
+* Create update service for books ([e68a32c](https://github.com/esgiraldop/library-nest-js/commit/e68a32c71a2c68b072c90c6faace937e4c7adc16))
+
+### [0.0.19](https://github.com/esgiraldop/library-nest-js/compare/v0.0.18...v0.0.19) (2024-10-01)
+
+### [0.0.18](https://github.com/esgiraldop/library-nest-js/compare/v0.0.17...v0.0.18) (2024-10-01)
+
+### [0.0.17](https://github.com/esgiraldop/library-nest-js/compare/v0.0.16...v0.0.17) (2024-10-01)
+
+### [0.0.16](https://github.com/esgiraldop/library-nest-js/compare/v0.0.15...v0.0.16) (2024-10-01)
+
+### [0.0.15](https://github.com/esgiraldop/library-nest-js/compare/v0.0.14...v0.0.15) (2024-10-01)
+
+
+### Features
+
+* **book:** create endpoint for getting a book by id ([5826cdc](https://github.com/esgiraldop/library-nest-js/commit/5826cdcc8ecce9907d8e4ef5945b3441ff288342))
+
+### [0.0.14](https://github.com/esgiraldop/library-nest-js/compare/v0.0.11...v0.0.14) (2024-10-01)
+
+### Features
+
+- **book:** create endpoint for getting a book by id ([5826cdc](https://github.com/esgiraldop/library-nest-js/commit/5826cdcc8ecce9907d8e4ef5945b3441ff288342))
+
+### [0.0.13](https://github.com/esgiraldop/library-nest-js/compare/v0.0.11...v0.0.13) (2024-10-01)
+
+### Features
+
+- **book:** create endpoint for getting a book by id ([5826cdc](https://github.com/esgiraldop/library-nest-js/commit/5826cdcc8ecce9907d8e4ef5945b3441ff288342))
+
+### [0.0.12](https://github.com/esgiraldop/library-nest-js/compare/v0.0.11...v0.0.12) (2024-10-01)
+
+### Features
+
+- **book:** create endpoint for getting a book by id ([5826cdc](https://github.com/esgiraldop/library-nest-js/commit/5826cdcc8ecce9907d8e4ef5945b3441ff288342))
+
 ### [0.0.11](https://github.com/esgiraldop/library-nest-js/compare/v0.0.10...v0.0.11) (2024-10-01)
 
 ### [0.0.10](https://github.com/esgiraldop/library-nest-js/compare/v0.0.9...v0.0.10) (2024-10-01)
@@ -14,17 +83,15 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### [0.0.6](https://github.com/esgiraldop/library-nest-js/compare/v0.0.5...v0.0.6) (2024-10-01)
 
-
 ### Features
 
-* **book:** implemented find all books ([2faea44](https://github.com/esgiraldop/library-nest-js/commit/2faea4444eaf1c6db06ca43144253c0013230438))
+- **book:** implemented find all books ([2faea44](https://github.com/esgiraldop/library-nest-js/commit/2faea4444eaf1c6db06ca43144253c0013230438))
 
 ### [0.0.5](https://github.com/esgiraldop/library-nest-js/compare/v0.0.4...v0.0.5) (2024-09-30)
 
-
 ### Features
 
-* **book:** implement database connection and creat book service ([01d56e6](https://github.com/esgiraldop/library-nest-js/commit/01d56e6d76ce40df44b3ded7199a39bc1322b4a8))
+- **book:** implement database connection and creat book service ([01d56e6](https://github.com/esgiraldop/library-nest-js/commit/01d56e6d76ce40df44b3ded7199a39bc1322b4a8))
 
 ### [0.0.4](https://github.com/esgiraldop/library-nest-js/compare/v0.0.3...v0.0.4) (2024-09-26)